+++ src/main/java/ru/topjava/lunchvoter/repository/CrudRestaurantRepository.java
@@ -5,6 +5,6 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.topjava.lunchvoter.model.Restaurant;
 
 @Transactional(readOnly = true)
-public interface CrudRestaurantRepository extends JpaRepository<Restaurant, Long> {
-    Restaurant getReferenceById(Long id);
+public interface CrudRestaurantRepository extends JpaRepository<Restaurant, Integer> {
+    Restaurant getReferenceById(Integer id);
 }
