+++ src/main/java/ru/topjava/lunchvoter/repository/RestaurantRepository.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Repository;
 import ru.topjava.lunchvoter.model.Restaurant;
 
 @Repository
-public class RestaurantRepository extends BaseRepository<Restaurant, Long> {
+public class RestaurantRepository extends BaseRepository<Restaurant, Integer> {
 
     private final CrudRestaurantRepository crudRestaurantRepository;
