+++ src/main/java/ru/topjava/lunchvoter/repository/VoteRepository.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Repository;
 import ru.topjava.lunchvoter.model.Vote;
 
 @Repository
-public class VoteRepository extends BaseRepository<Vote, Long> {
+public class VoteRepository extends BaseRepository<Vote, Integer> {
 
     private final CrudVoteRepository crudVoteRepository;
 
@@ -13,7 +13,7 @@ public class VoteRepository extends BaseRepository<Vote, Long> {
         this.crudVoteRepository = crudVoteRepository;
     }
 
-    public Vote findByUserId(Long userId) {
+    public Vote findByUserId(Integer userId) {
         return crudVoteRepository.findByUserId(userId);
     }
 }
